+++ DamPizzaApp/private/cache/retriever/catalog.xml
@@ -3,4 +3,5 @@
     <system systemId="http://javafx.com/javafx/8" uri="www.oracle.com/technetwork/java/javase/overview/index.html"/>
     <system systemId="http://javafx.com/javafx/9.0.1" uri="www.oracle.com/technetwork/java/javase/overview/index.html"/>
     <system systemId="http://javafx.com/fxml/1" uri="www.oracle.com/technetwork/java/javase/overview/index.html"/>
+    <system systemId="http://javafx.com/javafx/8.0.141" uri="www.oracle.com/technetwork/java/javase/overview/index.html"/>
 </catalog>
