+++ DamPizzaApp/src/main/java/com/dampizza/logic/imp/ProductManagerImp.java
@@ -153,13 +153,13 @@ public class ProductManagerImp implements ProductManagerInterface {
         try {
             tx = session.beginTransaction();
 
-            // Retrieve user to update
-            Query query = session.createQuery(hql);
-            query.setParameter("id", id);
-            ProductEntity productToDelete = (ProductEntity) query.uniqueResult();
+            // Retrieve user to delete
+
+            ProductEntity productToDelete = (ProductEntity) session.get(ProductEntity.class, id);
 
             if (productToDelete != null) {
                 session.delete(productToDelete);
+
                 res = 1;
                 logger.log(Level.INFO, "Product id<{0}>, name<{1}> deleted.", new Object[]{id, productToDelete.getName()});
             } else {
