+++ DamPizzaApp/src/main/java/com/dampizza/logic/io/OrderManagerInterface.java
@@ -10,12 +10,13 @@ import com.dampizza.exception.order.OrderDeleteException;
 import com.dampizza.exception.order.OrderQueryException;
 import com.dampizza.exception.order.OrderUpdateException;
 import com.dampizza.logic.dto.OrderDTO;
+import com.dampizza.logic.dto.UserDTO;
 import java.util.List;
 
 /**
  * Order Manager Interface.
  * CRUD methods for OrderEntity
- * @author Carlos
+ * @author Carlos Santos
  */
 public interface OrderManagerInterface {
     
@@ -44,7 +45,17 @@ public interface OrderManagerInterface {
     public Integer updateStatus(Long id, Integer status) throws OrderUpdateException, OrderQueryException;
     
     /**
-     * deleter order entity
+     * 
+     * @param id order id
+     * @param dealerId dealer id
+     * @return 1 success, 2 order not found, 0 error.
+     * @throws OrderUpdateException
+     * @throws OrderQueryException 
+     */
+    public Integer updateDealer(Long id, Long dealerId) throws OrderUpdateException, OrderQueryException;
+    
+    /**
+     * Delete order entity
      * @param id order to delete id
      * @return Success: 1, Order not found:2, Error:0
      * @throws OrderDeleteException
