+++ DamPizzaApp/src/main/java/com/dampizza/model/entity/Item.java
@@ -7,7 +7,7 @@ package com.dampizza.model.entity;
 
 /**
  * Item POJO
- * @author Carlos
+ * @author Carlos Santos
  */
 public class Item {
     private Integer qty;
