+++ DamPizzaApp/src/main/java/com/dampizza/views/custom/orderList.java
@@ -3,24 +3,24 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.dampizza.views.user.manager;
+package com.dampizza.views.custom;
 
-import com.dampizza.logic.dto.ProductDTO;
 import com.gluonhq.charm.glisten.control.CharmListCell;
 import com.gluonhq.charm.glisten.control.ListTile;
-
-
+import com.dampizza.logic.dto.OrderDTO;
+import javafx.beans.value.ObservableValue;
+import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 /**
  *
  * @author Jon Xabier Gimenez
  */
-public class productList extends CharmListCell<ProductDTO>{
+public class orderList extends CharmListCell<OrderDTO>{
     
     private final ListTile tile;
-      private final ImageView imageView;
+    private final ImageView imageView;
 
-    public productList() {
+    public orderList() {
         this.tile = new ListTile();
         imageView = new ImageView();
         tile.setPrimaryGraphic(imageView);
@@ -28,11 +28,11 @@ public class productList extends CharmListCell<ProductDTO>{
     }
 
     @Override
-    public void updateItem(ProductDTO item, boolean empty) {
+    public void updateItem(OrderDTO item, boolean empty) {
         super.updateItem(item, empty);
         if (item != null && !empty) {
             
-            tile.textProperty().setAll(item.getName()+"    "+item.getPrice(),item.getDescription());
+            tile.textProperty().setAll(item.getId()+"  "+item.getDate(),item.getAddress());
             setGraphic(tile);
         } else {
             setGraphic(null);
