+++ DamPizzaApp/src/main/java/com/dampizza/views/login/LoginView.java
@@ -10,7 +10,7 @@ import javafx.stage.Stage;
 /**
  * Login view provider (FXML Loader)
  *
- * @author Carlos Santos
+ * @author ???
  */
 public class LoginView {
