--- DamPizzaApp/src/main/java/com/dampizza/views/user/CustomerView.java
@@ -1,34 +1,0 @@
-package com.dampizza.views.user;
-
-import com.dampizza.views.login.*;
-import com.gluonhq.charm.glisten.mvc.View;
-import java.io.IOException;
-import javafx.fxml.FXMLLoader;
-
-/**
- * Customer view provider (FXML Loader)
- * @author Carlos Santos
- */
-public class CustomerView {
-
-    private final String name;
-
-    public CustomerView(String name) {
-        this.name = name;
-    }
-    
-    /**
-     * Load and returns fxml view.
-     * @return view object
-     */
-    public View getView() {
-        try {
-            View view = FXMLLoader.load(CustomerView.class.getResource("/com/dampizza/views/user/customer.fxml"));
-            view.setName(name);
-            return view;
-        } catch (IOException e) {
-            System.out.println("IOException: " + e);
-            return new View(name);
-        }
-    }
-}
