+++ DamPizzaApp/src/main/java/com/dampizza/views/user/customer/CustomerPresenter.java
@@ -1,4 +1,4 @@
-package com.dampizza.views.user;
+package com.dampizza.views.user.customer;
 
 import com.dampizza.views.login.*;
 import com.gluonhq.charm.glisten.application.MobileApplication;
