+++ DamPizzaApp/src/main/java/com/dampizza/views/user/dealer/DealerOrderPresenter.java
@@ -1,9 +1,5 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-import com.dampizza.views.user.manager.*;
+package com.dampizza.views.user.dealer;
+
 import com.dampizza.App;
 import static com.dampizza.App.DEALER_VIEW;
 
@@ -14,6 +10,7 @@ import com.dampizza.exception.order.OrderUpdateException;
 import com.dampizza.logic.dto.ProductDTO;
 import com.dampizza.logic.imp.OrderManagerImp;
 import com.dampizza.util.LogicFactory;
+import com.dampizza.views.custom.productList;
 import com.gluonhq.charm.glisten.application.MobileApplication;
 import com.gluonhq.charm.glisten.application.ViewStackPolicy;
 import com.gluonhq.charm.glisten.control.AppBar;
