+++ DamPizzaApp/src/main/java/com/dampizza/views/user/dealer/DealerOrderView.java
@@ -1,6 +1,5 @@
 package com.dampizza.views.user.manager.dealer;
 
-import com.dampizza.views.user.*;
 import com.dampizza.views.login.*;
 import com.gluonhq.charm.glisten.mvc.View;
 import java.io.IOException;
@@ -8,7 +7,7 @@ import javafx.fxml.FXMLLoader;
 
 /**
  * Customer view provider (FXML Loader)
- * @author Carlos Santos
+ * @author ???
  */
 public class DealerOrderView {
