+++ DamPizzaApp/src/main/java/com/dampizza/views/user/dealer/DealerPresenter.java
@@ -1,12 +1,5 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.dampizza.views.user.manager.dealer;
-
 import com.dampizza.util.LogicFactory;
-import com.dampizza.views.user.manager.orderList;
+import com.dampizza.views.custom.orderList;
 
 import com.gluonhq.charm.glisten.control.CharmListView;
