+++ DamPizzaApp/src/main/java/com/dampizza/views/user/manager/ManagerView.java
@@ -7,7 +7,7 @@ import javafx.fxml.FXMLLoader;
 
 /**
  * Manager view provider (FXML Loader)
- * @author Carlos Santos
+ * @author ???
  */
 public class ManagerView {
