+++ DamPizzaApp/src/main/resources/com/dampizza/views/user/customer/customer.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.VBox?>
 
-<View fx:id="primary" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="350.0" stylesheets="@Users/2dam/dampizza/DamPizzaApp/src/main/resources/com/dampizza/style.css" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.dampizza.views.user.CustomerPresenter">
+<View fx:id="primary" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="350.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.dampizza.views.user.customer.CustomerPresenter">
    <top>
       <VBox alignment="TOP_CENTER" prefHeight="102.0" prefWidth="350.0" BorderPane.alignment="CENTER" />
    </top>
